requirements.txt
# CyberKit Python Dependencies
# =============================

# Core dependencies (recommended for enhanced features)
# Install with: pip install -r requirements.txt

# YAML configuration support
pyyaml>=6.0

# Optional: Enhanced features
# Uncomment the ones you want to use

# API requests (for OSINT modules with API integrations)
# requests>=2.28.0

# Python-nmap wrapper (for programmatic nmap usage)
# python-nmap>=0.7.1

# Rich terminal output (for better UI)
# rich>=13.0.0

# Progress bars
# tqdm>=4.65.0

# ===========================
# System Tool Requirements
# ===========================
# CyberKit integrates with system security tools via subprocess
# Install these tools on Kali Linux:
#
# sudo apt update && sudo apt install -y \
#   nmap masscan \
#   gobuster ffuf dirb \
#   nikto sqlmap \
#   hydra medusa \
#   john hashcat \
#   metasploit-framework \
#   whatweb theharvester sherlock \
#   sslscan dnsrecon wpscan \
#   netcat-openbsd \
#   aircrack-ng \
#   enum4linux smbclient \
#   crackmapexec
#
# Optional tools:
#   bloodhound crackmapexec impacket-scripts
